Merge
Former-commit-id: 19bef9d0a0cf4335ee6c7fdca74d1a378e8c618c
Former-commit-id: 72d8ff32dd288d5ba43d0ce1f1c5ecd642d48721

+++ web/concrete/src/Workflow/BasicWorkflow.php
@@ -66,18 +66,24 @@ class BasicWorkflow extends \Concrete\Core\Workflow\Workflow
     {
         // lets save the basic data associated with this workflow.
         $req = $wp->getWorkflowRequestObject();
-        $db = Loader::db();
-        $db->Execute(
-            'INSERT INTO BasicWorkflowProgressData (wpID, uIDStarted) VALUES (?, ?)',
-            array($wp->getWorkflowProgressID(), $req->getRequesterUserID()));
+        
+        // Check if the workflow is not already approved
+        if (is_object($req)) {
+        
+            if ($this->canApproveWorkflow()) {
+                // Then that means we have the ability to approve the workflow we just started.
+                // In that case, we transparently approve it, and skip the entry notification step.
+                $wpr = $req->approve($wp);
+                $wp->delete();
 
-        if ($this->canApproveWorkflow()) {
-            // Then that means we have the ability to approve the workflow we just started.
-            // In that case, we transparently approve it, and skip the entry notification step.
-            $wpr = $req->approve($wp);
-            $wp->delete();
-        } else {
-            $ui = UserInfo::getByID($req->getRequesterUserID());
+            } else {
+
+                $db = Loader::db();
+                $db->Execute(
+                    'INSERT INTO BasicWorkflowProgressData (wpID, uIDStarted) VALUES (?, ?)',
+                    array($wp->getWorkflowProgressID(), $req->getRequesterUserID()));
+
+                $ui = UserInfo::getByID($req->getRequesterUserID());
 
             // let's get all the people who are set to be notified on entry
             $message = t(
